Event Driven LCD Shield Button Press
Allows the use of Events as opposed to a while loop in order to change
the LCD display based on user input.  In order for this to work an extra
wire must be soldered onto Pin 20 (INTA) of the MCP23017 chip and
connected to Digital I/O Pin 10 (as per the code) in order to read the
interrupt and get data as to the button pressed.