Fixed Button Press Issue Added ColorCount
Added a color count value to the BacklightColor enum so that looping is
made easier.
I had an issue with this when I coded the functionality to change the
background display because as soon as I changed the Background color
from the default of White (which  was set in the RGBLCDShield class
constructor) my buttons on the LCD shield would stop working. I
discovered that only the first 5 bits of the InterruptBits[0] byte
contained the button that was pressed.  What was happening was that when
I changed the LCD backlight color, additional bits were being added to
the front of the byte and so the integer that was cast with
(NetduinoRGBLCDShield.Button)InterruptBits[0] in my switch statement did
not recognize a button press...  I observed this as my UI stopping
working and had a LOT of difficulty troubleshooting! I added the below
default to the Switch statement so that future debugging sessions would
verify that my button events were firing, but I will still recieve a lot
of false posotives since a single button press will also send empty
bytes... At first I tried to shift the bits using
(NetduinoRGBLCDShield.Button)(InterruptBits[0] >> 3) which would shift
the left (if InterruptBits[0] was 1000, the statement would yield 62)
but realized that this wasn't the proper method since I was still
keeping the higher order bits. The solution was to perform a bitwise AND
operation (&) on the values using a mask that masked out the pertinant
bits (in this case 31 or 0x1F or 11111) so that if InterruptBits[0] was
set at 1,000 (1111101000 in binary) then an AND against 31 (11111
binary) would yield 8 (1111101000 AND 0000011111 = 1000)